This patch fixes the "Stop Scanning" button in the notification pulldown.
The old code had a secondary code path to disable the scanning in TurnOffReceiver.
This patch changes the TurnOffReceiver to just send an intent to disable the scanning using a local broadcast.

The MapActivity sent an ACTION_UNPAUSE_SCANNING intent which was being accepted by the MainApp, this has been moved into the MainActivity instead..

When MainActivity::onToggleScanningClickedView() is called, we force the UI to update.  This was previously causing problems where the MapActivity would enable the scan, but the button UI was not properly refreshing.